Major changes to get VRTK integrated into the game:
VRTK now loads SDK's and the MapLoader will wait for them to be loaded before proceeding with MapGeneration via a Coroutine
This is just another reason why we need to prioritize making MapGrid operations atomic

Renamed Grid to MapGrid finally due to conflicts with the Unity default Grid class. This was necessary to use some VRTK functions that related to Grid. Adjusted every reference to Grid in all of our code to now work with the MapGrid name.

Implemented a Controller button test class

Implemented a DPadController that allows for classic DPad 4-axis style controls on the Vive Touchpad

Implemented two new types of Controllers based on DPadController that uses VRTK code to generate map tiles automagically and move the player around the map, thusly returning the original function of the controllers that was once handled by the Player class

Adjusted the MapGenerator and MapLoader to make some checks to prevent NullObjectException at runtime

Removed the awful debug spam from the AI and State classes.

Pulled out the Grid coordinate to Worldspace coordinate converter helper function from MobileEntity to be part of MapGrid, since the MapGrid is the main authoritative source about Grid things. Might still be broken, however. Untested.

Player is now translated to the proper location on the map and does not need to duck using the new MapGrid grid coordinate converter above

Player is automatically set to the correct CameraRig depending on SDK load type

ControllerTouchObject deprecated and renamed to SteamVRControllerTouchObject, it's using SteamVR specific code, don't use it for the future until we can convert it to VRTK code