Tile Prefab Child Element Naming Convention Changes:
I've decided to change the way we'll be naming Children GameObjects that are part of a Tile master parent prefab. This means the Walls, Cielings, and Floors of the prefab for now. We used to specify them with the Model name, or possibly the Tile name then a '-' character and finally the type of Child it was (e.g., "Wall S" for the southern wall) so that the total Child GameObject name would be something like "SciFi-Wall S" but I feel like this is unnecessary, the type of tile is clearly defined by the parent, and the original wall prefabs would be renamed anyways with their cardinal direction, so it just makes it easier from a code standpoint (we don't need to include the Tile name in our removeType() calls anymore) if we just name them according to what type they are and that's it (like "Ceiling" instead of "SciFi-Ceiling").

So soon I'll be reuploading all of our prefabs to respect this naming convention. My next commit will be a test of overwriting the SciFi Prefabs using the tool. This means we'll lose some of the nice additions (lights/barrels/etc) I made, but I'll just add them back.