Major change to gridobject ID numbers! (currently known as blocks.key.txt)
In order to get the TilePrefabGenerator working, I need to load in master tiles for the floor and ceiling type Tiles, since ceiling corners and such have different wall rotations.
Currently we stored ceiling master prefabs in ID 20, but this was just because I was doing them by hand and ended up doing them last.
Since the ID 1 is completely useless, I decided to change ID 1 into ID 20 and remove ID 20 so it is now properly specified in the gridobject list. This doesn't affect level generation as it is now, since nothing uses ID 1.
The blocks.key.txt has been updated to reflect these changes. I will soon be changing the name of that file, because it determines the ID's of both Blocks and Tiles, and any GridObject really, it's just a relic of old programming that it has that name

To create a new Tile prefab to give to the generator, you can create a 0 ID prefab with the floor-type configuration. If you don't want to create a ceiling variation, you can just duplicate that prefab over to prefab ID 1. It will work as expected.
The TilePrefabGenerator has been updated to intelligently decide on whether or not it should use the Floor or Ceiling master prefab as its base for generating other prefabs, based on their IDs and seems to do so correctly as of this point.