Finished SciFiTile-Prefab Generation update:
The base of these prefabs was computed by the new generator, from the two master tiles 00 and 01.

The masters have been updated to have columns, which prevent the holes in the map we were having and look prettier. They'll look even better when we get scale correct.

The columns have a collider which may not be necessary, but is there for completeness, and they have been appropriately scaled so that the ceilings and floors meet up right and with minimal z-fighting.

The new SciFi-Pillar tile component has been created.

Moved Turret to it's own prefab, so it's not sitting around in the scene.

Adding some basic lighting back to the tile prefabs.

Tile prefab generator now sets the prefabs to active so you can see the changes in the resources folder in the unity inspector, allowing you to quickly look over prefab changes.