Created the EntityPart class, and more:
The EntityPart class lets us define sub-parts of Entities that absorb damage differently than the main body.
It can be used to create critical hit zones, or armor, or breakaway pieces.

I've installed these on the TreadBot, giving the TreadBot working armor for the front and sides, as well as a weak point: it's Fuel Tank.

Two shots to the fuel tank knocks it out easily, but you'll have to get behind it.

The gun is a breakaway item that persists. Once you shoot the turret off, you get rewarded with the Turret to use as a weapon!
That's right, you can literally rip off a turret from a TreadBot and shoot it as a weapon, which is so freaking cool.

All it needs is some tweaks to test and polish it, such as fix the orientation of the dropped weapon, and etc. This can't really be done until we get half-sized tiles working.