Added StateMachine capabilities using Mecanim, unfinished TreadBot states
The State Machine code allows us to create FSM for Entities, allowing us to streamline their behavioural coding

Currently I am doing all of the actual logic in new function inside the Entity class, but this is not ideal, as the StateMachineController doesn't have much control and it kind of defeats the point for us to be using the Mecanim FSM if all our logic is being done in Entity anyways. Especially considering it actually takes MORE code to use this FSM system than if we were to code it manually.

The TreadBot has been updated to use the new FSM code and it technically works, but it's states oscillate VERY rapidly once it starts shooting, and this bug will NEED to be fixed before we build it again. Otherwise it could be a performance issue.

In general, the State Machine code for the TreadBot in it's new Event method needs to be thoroughly vetted. But it at least functions right now.

We'll need to move Drone over to using this system later, and of course remove the old FSM code that we never ended up really using.