Made GridObjects Configuration ID numbers into an enum
This makes it easier to check if a number is a valid GridObject Configuration ID Number. If we support the number, the Enum.IsDefined method can tell us if it's in our enum or not.
The other advantage is easier readability in code, especially for the TilePrefabGenerator thing which has tons of ID definitions to check, which previously required cross-checking with the blocks.key.txt file
The blocks.key.txt file has been updated to reflect the changes, and you should not use it anymore, since we now use the enum as our official source for GridObjectConfiguration numbers.
We may need to move GridObjectConfiguration enum into Tile and call it a TileConfiguration because I have trouble understanding how walls and floors are relevant to Blocks, but that's for another day.

NOTE: I wanted to also perform this enumeration to the GridSpaceConfiguration number as well. unfortunately, since GridSpaceConfigurations are defined by an XML flatfile, we can't use an enum since they're set dynamically upon read of that file
This is okay, we might just have to use a dictionary or something. But we can't get around using XML flatfile there because the numbers also have the GridObjects Configurations associated of them (8 to be exact) and that can't be stored with an enum.
It's not a huge issue, it just might confuse people to see that Configurations are stored as an int in GridSpace but an enum in GridObject. They're both configurations, still, however.

Next up is to start implementing the new halfsized tiles into the MapGenerator code, and I suppose, the Tile Prefab Generator if we haven't done that already...